+++ pkg/ruler/ruler_test.go
@@ -301,7 +301,7 @@ func TestSharding(t *testing.T) {
 
 		"default sharding, single ruler": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyDefault,
+			shardingStrategy: util.ShardingStrategyDefault,
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", []uint32{0}, ring.ACTIVE, time.Now())
 			},
@@ -310,7 +310,7 @@ func TestSharding(t *testing.T) {
 
 		"default sharding, multiple ACTIVE rulers": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyDefault,
+			shardingStrategy: util.ShardingStrategyDefault,
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", sortTokens([]uint32{user1Group1Token + 1, user2Group1Token + 1}), ring.ACTIVE, time.Now())
 				desc.AddIngester(ruler2, ruler2Addr, "", sortTokens([]uint32{user1Group2Token + 1, user3Group1Token + 1}), ring.ACTIVE, time.Now())
@@ -331,7 +331,7 @@ func TestSharding(t *testing.T) {
 
 		"default sharding, unhealthy ACTIVE ruler": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyDefault,
+			shardingStrategy: util.ShardingStrategyDefault,
 
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", sortTokens([]uint32{user1Group1Token + 1, user2Group1Token + 1}), ring.ACTIVE, time.Now())
@@ -355,7 +355,7 @@ func TestSharding(t *testing.T) {
 
 		"default sharding, LEAVING ruler": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyDefault,
+			shardingStrategy: util.ShardingStrategyDefault,
 
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", sortTokens([]uint32{user1Group1Token + 1, user2Group1Token + 1}), ring.LEAVING, time.Now())
@@ -371,7 +371,7 @@ func TestSharding(t *testing.T) {
 
 		"default sharding, JOINING ruler": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyDefault,
+			shardingStrategy: util.ShardingStrategyDefault,
 
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", sortTokens([]uint32{user1Group1Token + 1, user2Group1Token + 1}), ring.JOINING, time.Now())
@@ -387,7 +387,7 @@ func TestSharding(t *testing.T) {
 
 		"shuffle sharding, single ruler": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 
 			setupRing: func(desc *ring.Desc) {
 				desc.AddIngester(ruler1, ruler1Addr, "", sortTokens([]uint32{0}), ring.ACTIVE, time.Now())
@@ -400,7 +400,7 @@ func TestSharding(t *testing.T) {
 
 		"shuffle sharding, multiple rulers, shard size 1": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 			shuffleShardSize: 1,
 
 			setupRing: func(desc *ring.Desc) {
@@ -417,7 +417,7 @@ func TestSharding(t *testing.T) {
 		// Same test as previous one, but with shard size=2. Second ruler gets all the rules.
 		"shuffle sharding, two rulers, shard size 2": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 			shuffleShardSize: 2,
 
 			setupRing: func(desc *ring.Desc) {
@@ -434,7 +434,7 @@ func TestSharding(t *testing.T) {
 
 		"shuffle sharding, two rulers, shard size 1, distributed users": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 			shuffleShardSize: 1,
 
 			setupRing: func(desc *ring.Desc) {
@@ -454,7 +454,7 @@ func TestSharding(t *testing.T) {
 		},
 		"shuffle sharding, three rulers, shard size 2": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 			shuffleShardSize: 2,
 
 			setupRing: func(desc *ring.Desc) {
@@ -478,7 +478,7 @@ func TestSharding(t *testing.T) {
 		},
 		"shuffle sharding, three rulers, shard size 2, ruler2 has no users": {
 			sharding:         true,
-			shardingStrategy: ShardingStrategyShuffle,
+			shardingStrategy: util.ShardingStrategyShuffle,
 			shuffleShardSize: 2,
 
 			setupRing: func(desc *ring.Desc) {