+++ pkg/storegateway/gateway_test.go
@@ -58,14 +58,14 @@ func TestConfig_Validate(t *testing.T) {
 		"should fail if the sharding strategy is shuffle-sharding and shard size has not been set": {
 			setup: func(cfg *Config, limits *validation.Limits) {
 				cfg.ShardingEnabled = true
-				cfg.ShardingStrategy = ShardingStrategyShuffle
+				cfg.ShardingStrategy = util.ShardingStrategyShuffle
 			},
 			expected: errInvalidTenantShardSize,
 		},
 		"should pass if the sharding strategy is shuffle-sharding and shard size has been set": {
 			setup: func(cfg *Config, limits *validation.Limits) {
 				cfg.ShardingEnabled = true
-				cfg.ShardingStrategy = ShardingStrategyShuffle
+				cfg.ShardingStrategy = util.ShardingStrategyShuffle
 				limits.StoreGatewayTenantShardSize = 3
 			},
 			expected: nil,
@@ -244,45 +244,45 @@ func TestStoreGateway_BlocksSharding(t *testing.T) {
 			expectedBlocksLoaded: 2 * numBlocks, // each gateway loads all the blocks
 		},
 		"default sharding strategy, 1 gateway, RF = 1": {
-			shardingStrategy:     ShardingStrategyDefault,
+			shardingStrategy:     util.ShardingStrategyDefault,
 			replicationFactor:    1,
 			numGateways:          1,
 			expectedBlocksLoaded: numBlocks,
 		},
 		"default sharding strategy, 2 gateways, RF = 1": {
-			shardingStrategy:     ShardingStrategyDefault,
+			shardingStrategy:     util.ShardingStrategyDefault,
 			replicationFactor:    1,
 			numGateways:          2,
 			expectedBlocksLoaded: numBlocks, // blocks are sharded across gateways
 		},
 		"default sharding strategy, 3 gateways, RF = 2": {
-			shardingStrategy:     ShardingStrategyDefault,
+			shardingStrategy:     util.ShardingStrategyDefault,
 			replicationFactor:    2,
 			numGateways:          3,
 			expectedBlocksLoaded: 2 * numBlocks, // blocks are replicated 2 times
 		},
 		"default sharding strategy, 5 gateways, RF = 3": {
-			shardingStrategy:     ShardingStrategyDefault,
+			shardingStrategy:     util.ShardingStrategyDefault,
 			replicationFactor:    3,
 			numGateways:          5,
 			expectedBlocksLoaded: 3 * numBlocks, // blocks are replicated 3 times
 		},
 		"shuffle sharding strategy, 1 gateway, RF = 1, SS = 1": {
-			shardingStrategy:     ShardingStrategyShuffle,
+			shardingStrategy:     util.ShardingStrategyShuffle,
 			tenantShardSize:      1,
 			replicationFactor:    1,
 			numGateways:          1,
 			expectedBlocksLoaded: numBlocks,
 		},
 		"shuffle sharding strategy, 5 gateways, RF = 2, SS = 3": {
-			shardingStrategy:     ShardingStrategyShuffle,
+			shardingStrategy:     util.ShardingStrategyShuffle,
 			tenantShardSize:      3,
 			replicationFactor:    2,
 			numGateways:          5,
 			expectedBlocksLoaded: 2 * numBlocks, // blocks are replicated 2 times
 		},
 		"shuffle sharding strategy, 20 gateways, RF = 3, SS = 3": {
-			shardingStrategy:     ShardingStrategyShuffle,
+			shardingStrategy:     util.ShardingStrategyShuffle,
 			tenantShardSize:      3,
 			replicationFactor:    3,
 			numGateways:          20,
@@ -360,7 +360,7 @@ func TestStoreGateway_BlocksSharding(t *testing.T) {
 			assert.Equal(t, float64(testData.expectedBlocksLoaded), metrics.GetSumOfGauges("cortex_bucket_store_blocks_loaded"))
 			assert.Equal(t, float64(2*testData.numGateways), metrics.GetSumOfGauges("cortex_bucket_stores_tenants_discovered"))
 
-			if testData.shardingStrategy == ShardingStrategyShuffle {
+			if testData.shardingStrategy == util.ShardingStrategyShuffle {
 				assert.Equal(t, float64(testData.tenantShardSize*numBlocks), metrics.GetSumOfGauges("cortex_blocks_meta_synced"))
 				assert.Equal(t, float64(testData.tenantShardSize*numUsers), metrics.GetSumOfGauges("cortex_bucket_stores_tenants_synced"))
 			} else {